Detect if server was already terminated

+++ language_tool_python/server.py
@@ -305,6 +305,8 @@ class LanguageTool:
 
     def _terminate_server(self):
         LanguageToolError_message = ''
+        if self._server is None:
+            return
         try:
             self._server.terminate()
         except OSError:
